+++ app/src/services/logEvents.ts
@@ -849,13 +849,6 @@ const _deprecatedLogIndicatorObdLvl2 = async (subThemes, count) => {
       value: subThemes[i],
     });
   }
-  // Also log the count
-  await logEvent({
-    category: ONBOARDING,
-    action: INDICATOR_OBD_LVL2,
-    name: "count",
-    value: count,
-  });
 };
 
 const _deprecatedLogIndicatorObdValidate = async (indicators: number[]) => {
@@ -1363,6 +1356,24 @@ const logOpenIndicatorsSettings = async () => {
   });
 };
 
+const logViewedArticlesList = async (nbArticlesDisplayed: number) => {
+  await logEvent({
+    category: "RESOURCES",
+    action: "VIEWED_ARTICLES_LIST",
+    name: "nb_articles_displayed",
+    value: nbArticlesDisplayed,
+  });
+};
+
+const logSelectedCategory = async (categoryId: number) => {
+  await logEvent({
+    category: "RESOURCES",
+    action: "SELECTED_CATEGORY",
+    name: "category_id",
+    value: categoryId,
+  });
+};
+
 export default {
   initMatomo,
   logAppVisit,
@@ -1518,4 +1529,7 @@ export default {
   logDeleteDrug,
   logToggleDrug,
   logOpenDrugSettings,
+  // Resources events
+  logViewedArticlesList,
+  logSelectedCategory,
 };